split tau2 implementation into taskset and harness
Convert the flat tau2_bench_v1.py module into a package with taskset.py
(taskset, task, config, reward, data bootstrap) and harness.py (harness,
config, subprocess runner, LiteLLM patches).
