+++ environments/tau2_bench_v1/pyproject.toml
@@ -14,8 +14,8 @@ dependencies = [
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
-[tool.hatch.build]
-include = ["tau2_bench_v1.py", "pyproject.toml"]
+[tool.hatch.build.targets.wheel]
+packages = ["tau2_bench_v1"]
 
 [tool.hatch.metadata]
 allow-direct-references = true
