+++ environments/tau2_bench_v1/tau2_bench_v1/__init__.py
@@ -1,0 +1,16 @@
+from tau2_bench_v1.harness import Tau2Harness, Tau2HarnessConfig
+from tau2_bench_v1.taskset import (
+    Tau2Domain,
+    Tau2Task,
+    Tau2Taskset,
+    Tau2TasksetConfig,
+)
+
+__all__ = [
+    "Tau2Domain",
+    "Tau2Harness",
+    "Tau2HarnessConfig",
+    "Tau2Task",
+    "Tau2Taskset",
+    "Tau2TasksetConfig",
+]
